third pass refactor lingering 200+ LOC files that have clear seams
